Misc. fixes and test instrumentation.

+++ imhotep-server/src/main/native/optimized_ftgs/optimized/src/splitter.cpp
@@ -7,7 +7,7 @@ namespace imhotep {
         // os.write(reinterpret_cast<const char *>(&term), sizeof(IntTerm));
         const int64_t  id(term.id());
         const int64_t doc_offset(term.doc_offset());
-        const int64_t doc_freq(term.doc_freq());
+        const int32_t doc_freq(term.doc_freq());
         os.write(reinterpret_cast<const char*>(&id), sizeof(id));
         os.write(reinterpret_cast<const char*>(&doc_offset), sizeof(doc_offset));
         os.write(reinterpret_cast<const char*>(&doc_freq), sizeof(doc_freq));
@@ -18,7 +18,7 @@ namespace imhotep {
         const std::string& id(term.id());
         const size_t       id_size(id.size());
         const int64_t     doc_offset(term.doc_offset());
-        const int64_t     doc_freq(term.doc_freq());
+        const int32_t     doc_freq(term.doc_freq());
         os.write(reinterpret_cast<const char*>(&id_size), sizeof(id_size));
         os.write(id.data(), id.size());
         os.write(reinterpret_cast<const char*>(&doc_offset), sizeof(doc_offset));
